+++ Sources/MongoSwift/Operations/CreateIndexesOperation.swift
@@ -3,13 +3,13 @@ import CLibMongoC
 /// Options to use when creating a new index on a `MongoCollection`.
 public struct CreateIndexOptions: Encodable {
     /// The maximum amount of time to allow the query to run - enforced server-side.
-    public var maxTimeMS: Int64?
+    public var maxTimeMS: Int?
 
     /// An optional `WriteConcern` to use for the command.
     public var writeConcern: WriteConcern?
 
     /// Initializer allowing any/all parameters to be omitted.
-    public init(maxTimeMS: Int64? = nil, writeConcern: WriteConcern? = nil) {
+    public init(maxTimeMS: Int? = nil, writeConcern: WriteConcern? = nil) {
         self.maxTimeMS = maxTimeMS
         self.writeConcern = writeConcern
     }
